Removed game over teleport delay
I removed the brief delay that occurred before teleporting users back to the lobby after a game ended. Initially, this delay was in place to ensure the game over sound played properly. I restructured the code so that the sound now plays after the teleportation, preventing any interruption of the sound without necessitating a delay in the teleportation process.