Max players per game config option
Previously, no cap was set for the
amount of players in one game zone.
There is now a cap set, it is 1 by
default but can be changed via the
config.yml. This cap ignores players
that are not in a snake game but may
still be within the game zone. This is
to prevent flagging users in spectator
mode or invisible admins.