Temp removal of PvP (see details)
The PvP gamemode was unfinished and had some glaring issues with the few parts that were done. I have all PvP related code in a backup and have decided to remove it for now to make code cleanup and the new region code merge easier. The gamemode update will likely be pushed back.